REF/TST: misplaced Categorical tests (#37678)
