+++ pandas/tests/arrays/categorical/test_analytics.py
@@ -359,10 +359,3 @@ class TestCategoricalAnalytics:
 
         with pytest.raises(ValueError, match=msg):
             cat.sort_values(inplace=value)
-
-    def test_isna(self):
-        exp = np.array([False, False, True])
-        c = Categorical(["a", "b", np.nan])
-        res = c.isna()
-
-        tm.assert_numpy_array_equal(res, exp)
